Move env helper to its own package
